requirements.txt
uvicorn
fastapi
aiohttp
asyncio
shortuuid
python-dotenv
sqlalchemy

pydantic
asyncpg
psycopg2-binary

celery[redis]
gevent

locust

